add route plan

+++ TASK/beep_task.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 extern uint8_t beep_duty;
-extern uint8_t beep_cycle;
+extern uint16_t beep_cycle;
 extern uint8_t beep_time;
 
 void beep_task_motor_stop(void);
